fix: add exception for rating field in RatingCoachAPIView

+++ app_run/models.py
@@ -112,12 +112,12 @@ class Subscripe(models.Model):
 
 
 class Rating(models.Model):
-    RATING_CHOICES = [(1, 1), (2, 2), (3, 3), (4, 4), (5, 5)]
+    RATING_CHOICES = [(1.0, 1), (2.0, 2), (3.0, 3), (4.0, 4), (5.0, 5)]
 
     athlete = models.ForeignKey(User, on_delete=models.CASCADE)
     coach = models.ForeignKey(User, on_delete=models.CASCADE, related_name="ratings")
 
-    rating = models.IntegerField(choices=RATING_CHOICES, null=True, blank=True)
+    rating = models.FloatField(choices=RATING_CHOICES, null=True, blank=True)
 
     def __str__(self):
         return f"{self.athlete.username} поставил тренеру {self.coach.username} рейтинг {self.rating}"
